requirements.txt
Django>=5.2,<6.0
Pillow
django-filter
requests
social-auth-app-django
Markdown>=3.4
bleach>=6.0